+++ src/main/java/org/team157/robot/BuildConstants.java
@@ -5,13 +5,13 @@ public final class BuildConstants {
     public static final String MAVEN_GROUP = "";
     public static final String MAVEN_NAME = "FRC2026";
     public static final String VERSION = "unspecified";
-    public static final int GIT_REVISION = 365;
-    public static final String GIT_SHA = "c36018c0b2bc10034e7244ede4d2e4dcddb103e4";
-    public static final String GIT_DATE = "2026-05-13 14:13:09 EDT";
+    public static final int GIT_REVISION = 366;
+    public static final String GIT_SHA = "2bdd41328abdfc8999eab7d5cad6e80f4c55241c";
+    public static final String GIT_DATE = "2026-05-15 11:56:30 EDT";
     public static final String GIT_BRANCH = "task/reexport-as-model";
-    public static final String BUILD_DATE = "2026-05-15 11:49:13 EDT";
-    public static final long BUILD_UNIX_TIME = 1778860153213L;
-    public static final int DIRTY = 1;
+    public static final String BUILD_DATE = "2026-05-15 15:57:29 EDT";
+    public static final long BUILD_UNIX_TIME = 1778875049705L;
+    public static final int DIRTY = 0;
 
     private BuildConstants() {}
 }
