+++ src/main/java/org/team157/robot/subsystems/drive/Drive.java
@@ -48,7 +48,7 @@ import org.team157.robot.Constants;
 import org.team157.robot.Constants.FieldConstants;
 import org.team157.robot.Constants.Mode;
 import org.team157.robot.generated.TunerConstants;
-import org.team157.robot.subsystems.SunstoneV2Mechanism3D.Mechanism3DConstants;
+import org.team157.robot.subsystems.SunstoneMechanism3D.Mechanism3DConstants;
 import org.team157.robot.util.LocalADStarAK;
 
 public class Drive extends SubsystemBase {
